chore: changed id related ops from int to long

+++ src/main/java/com/homelib/utils/BookInputReader.java
@@ -8,8 +8,8 @@ import java.util.Scanner;
 
 public interface BookInputReader {
     Book readBook();
-    int readId();
+    long readId();
     String listBooks();
-    int readIdDelete();
-    Long readIdUpdate();
+    long readIdDelete();
+    long readIdUpdate();
 }
